be-fastapi/pyproject.toml
[project]
name = "be-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.14.*",
    "asyncpg==0.30.*",
    "fastapi==0.115.*",
    "jinja2==3.1.*",
    "openai==1.61.*",
    "pdf2image==1.17.*",
    "pdfplumber==0.11.*",
    "pgvector==0.3.*",
    "pydantic-settings==2.7.*",
    "pytesseract==0.3.*",
    "python-docx==1.1.*",
    "python-multipart>=0.0.22",
    "python-pptx==1.0.*",
    "sqlalchemy[asyncio]==2.0.*",
    "sse-starlette==2.2.*",
    "tiktoken==0.8.*",
    "uvicorn[standard]==0.34.*",
    "weasyprint==63.*",
]

[dependency-groups]
dev = [
    "mypy==1.14.*",
    "pytest==8.*",
    "pytest-asyncio==0.25.*",
    "ruff==0.9.*",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
